requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
python-dotenv==1.0.0
pydantic==2.5.0
regex==2023.10.3
pytest==7.4.3
pytest-asyncio==0.21.1